error set

+++ app/Http/Livewire/University/Components/UploadLogo.php
@@ -36,8 +36,11 @@ class UploadLogo extends Component
 
     public function removeLogo(){
         $university = \Auth::user()->selected_university;
-        \Storage::disk('s3')->delete($university->monogram);
-        $university->update(['logo'=>null]);
+        if($university->monogram){
+            \Storage::disk('s3')->delete($university->monogram);  
+            $university->update(['logo'=>null]);  
+        }
+        
     }
